Rename style to styleText
